+++ test/integ_tests/test_device_creation.py
@@ -10,6 +10,7 @@
 # distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF
 # ANY KIND, either express or implied. See the License for the specific
 # language governing permissions and limitations under the License.
+
 from typing import List, Set
 
 import pytest
@@ -131,7 +132,6 @@ def test_device_enum():
     # validate all devices in enum
     providers = [getattr(Devices, attr) for attr in dir(Devices) if not attr.startswith("__")]
     for provider in providers:
-        devices = [getattr(provider, attr) for attr in dir(provider) if not attr.startswith("__")]
-        for arn in devices:
-            device = AwsDevice(arn)
+        for device_arn in provider:
+            device = AwsDevice(device_arn)
             _validate_device(device, active_providers)