+++ CHANGELOG.md
@@ -1,5 +1,23 @@
 # Changelog
 
+## v1.53.3 (2023-08-08)
+
+### Bug Fixes and Other Changes
+
+ * fix a bug in time series and add trapezoidal time series
+
+## v1.53.2 (2023-08-07)
+
+### Bug Fixes and Other Changes
+
+ * don't wrap FreeParameterExpression input as string
+
+## v1.53.1 (2023-08-03)
+
+### Bug Fixes and Other Changes
+
+ * Support OpenQASM `Program`s in `from_ir`
+
 ## v1.53.0.post0 (2023-08-02)
 
 ### Documentation Changes